refactor paths file structure
